@@ -14,7 +14,6 @@ classifiers = [
   "License :: OSI Approved :: Apache Software License",
   "Natural Language :: English",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
@@ -28,7 +27,7 @@ exclude = ["*.pyc", "*~", "*pycache*"]
 "CLI & Tools" = "https://github.com/johnpreston/cdk-gw-tools"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 requests = "^2.31"
 pyjwt = "^2.8"
 
@@ -39,6 +38,13 @@ isort = "^5.12"
 pyupgrade = "^3.3.1"
 pydantic = "^1.10.5"
 tbump = "^6.9.0"
+
+[tool.poetry.group.testing.dependencies]
+coverage = "^7.4.4"
+pytest = "^8.1.1"
+testcontainers = "^4.3.2"
+confluent-kafka = "^2.3.0"
+
 
 [tool.tbump]
 github_url = "https://codeberg.org/JohnPreston/cdk-proxy-api-client"
