@@ -1,5 +1,5 @@
 #   SPDX-License-Identifier: Apache-2.0
-#   Copyright 2023 John Mille <john@ews-network.net>
+#   Copyright 2024 John Mille <john@ews-network.net>
 
 from __future__ import annotations
 
@@ -17,10 +17,10 @@ from cdk_proxy_api_client.exceptions import (
 from cdk_proxy_api_client.proxy_api import ApiApplication
 
 
-class VirturalClusters(ApiApplication):
+class VirtualClusters(ApiApplication):
     app_path: str = "admin/vclusters"
 
-    def list_vclusters(self, as_list: bool = False) -> Union[Response, list[str]]:
+    def list_vclusters(self, as_list: bool = False) -> Response | list[str]:
         _path: str = f"{self.base_path}/"
         LOG.debug(f"list_vclusters path {_path}")
         req = self.proxy.client.get(_path, headers={"Accept": "application/json"})
@@ -34,7 +34,7 @@ class VirturalClusters(ApiApplication):
         username: str = None,
         lifetime_in_seconds: int = 86400,
         token_only: bool = False,
-    ) -> Union[Response, str]:
+    ) -> Response | str:
         """
         Docs: https://developers.conduktor.io/#tag/Virtual-Clusters/operation/Auth_v1_createClusterAccountToken
         Path: /admin/vclusters/v1/vcluster/{vcluster}/username/{username}
@@ -51,40 +51,96 @@ class VirturalClusters(ApiApplication):
             return req.json()["token"]
         return req
 
+    def create_concentration_rule(
+        self,
+        vcluster_name: str,
+        pattern: str,
+        delete_topic_name: str,
+        compact_topic_name: str = None,
+        compact_delete_topic_name: str = None,
+        cluster_id: str = None,
+    ) -> Response:
+        """
+        Docs: https://developers.conduktor.io/#tag/Virtual-Clusters/operation/ConcentrationRule_createConcentrationRule
+        Path: /admin/vclusters/v1/vcluster/{vcluster}/concentration-rules
+        """
+        _path: str = f"{self.base_path}/vcluster/{vcluster_name}/concentration-rules"
+        LOG.debug(f"create_concentration_rule path {_path}")
+        payload: dict = {
+            "pattern": pattern,
+            "physicalTopicName": delete_topic_name,
+            "physicalTopicCompactedName": compact_topic_name
+            if compact_topic_name
+            else f"{delete_topic_name}_compacted",
+            "physicalTopicCompactedDeletedName": compact_delete_topic_name
+            if compact_delete_topic_name
+            else f"{delete_topic_name}_compact_delete",
+        }
+        if cluster_id:
+            payload["clusterId"] = cluster_id
+        req = self.proxy.client.post(
+            _path, headers={"Accept": "application/json"}, json=payload
+        )
+        return req
+
+    def get_concentration_rules(self, vcluster_name: str) -> Response:
+        """
+        Docs: https://developers.conduktor.io/#tag/Virtual-Clusters/operation/ConcentrationRule_getConcentrationRules
+        Path: /admin/vclusters/v1/vcluster/{vcluster}/concentration-rules
+        """
+        _path: str = f"{self.base_path}/vcluster/{vcluster_name}/concentration-rules"
+        LOG.debug(f"get_concentration_rules path {_path}")
+        req = self.proxy.client.get(
+            _path,
+        )
+        return req
+
+    def delete_concentration_rule(self, vcluster_name: str) -> Response:
+        """
+        Docs: https://developers.conduktor.io/#tag/Virtual-Clusters/operation/ConcentrationRule_deleteConcentrationRule
+        Path: /admin/vclusters/v1/vcluster/{vcluster}/concentration-rules
+        """
+        _path: str = f"{self.base_path}/vcluster/{vcluster_name}/concentration-rules"
+        LOG.debug(f"delete_concentration_rule path {_path}")
+        req = self.proxy.client.delete(
+            _path,
+        )
+        return req
+
     def create_vcluster_topic_mapping(
         self,
         vcluster: str,
         logical_topic_name: str,
         physical_topic_name: str,
+        mapping_type: str = "alias",
         read_only: bool = False,
-        concentrated: bool = False,
         cluster_id: str = None,
     ) -> Response:
         """
         Docs: https://developers.conduktor.io/#tag/Virtual-Clusters/operation/Clusters_v1_createClusterTopicMapping
         Path: /admin/vclusters/v1/vcluster/{vcluster}/topics/{logicalTopicName}
         """
+        if mapping_type not in ["alias", "concentrated"]:
+            raise ValueError("mapping_type must be alias or concentrated")
+        if mapping_type == "concentrated":
+            raise ValueError("Must use concentration rules functions.")
         _path: str = (
             f"{self.base_path}/vcluster/{vcluster}/topics/{quote(logical_topic_name)}"
         )
         payload: dict = {
             "physicalTopicName": physical_topic_name,
             "readOnly": read_only,
-            "concentrated": concentrated,
+            "type": mapping_type,
         }
         if cluster_id:
             payload["clusterId"] = cluster_id
-        if payload["concentrated"] and payload["readOnly"]:
-            raise ValueError(
-                f"{_path} - concentrated is true, read_only cannot be true."
-            )
         LOG.debug(f"create_vcluster_topic_mapping path: {_path}")
         req = self.proxy.client.post(_path, json=payload)
         return req
 
     def list_vcluster_topic_mappings(
         self, vcluster: str, as_list: bool = False
-    ) -> Union[Response, list]:
+    ) -> Response | list[dict]:
         """
         Docs: https://developers.conduktor.io/#tag/Virtual-Clusters/operation/Clusters_v1_listClusterTopicMapping
         Path: /admin/vclusters/v1/vcluster/{vcluster}/topics
@@ -96,7 +152,7 @@ class VirturalClusters(ApiApplication):
             if as_list:
                 return req.json()
             return req
-        except GenericNotFound as error:
+        except GenericNotFound:
             raise VirtualClusterNotFound(vcluster)
 
     def delete_vcluster_topics_mappings(self, vcluster: str) -> Response:
