@@ -1,5 +1,5 @@
 #  SPDX-License-Identifier: Apache-2.0
-#  Copyright 2023 John Mille <john@ews-network.net>
+#  Copyright 2024 John Mille <john@ews-network.net>
 
 from requests import exceptions as req_exceptions
 
